% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcalculate_risk.R
\name{backcalcHelper_infectionRisk_perDay}
\alias{backcalcHelper_infectionRisk_perDay}
\title{Backcalculate risk: infection (helper function)}
\usage{
backcalcHelper_infectionRisk_perDay(infectionRisk_perDay, exposure_daysPerYear,
  target_infectionRisk_perYear = 1/10000)
}
\arguments{
\item{infectionRisk_perDay}{no default. will be set by function backcalc_infectionRisk_perDay()}

\item{exposure_daysPerYear}{exposure days per year}

\item{target_infectionRisk_perYear}{target infection risk per per year (default: 
1/10000)}
}
\value{
absolute offset between "target_infectionRisk_perYear" and 
"infectionRisk_perYear"
}
\description{
Backcalculate risk: infection (helper function)
}
\keyword{internal}
