% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{config_dummy_json}
\alias{config_dummy_json}
\title{config_default_json: default configuration developed by Christoph Sprenger 
for kwb.qmra::opencpu_simulate_risk() in JSON format}
\format{
A json list with all configuration parameters used by \link{opencpu_simulate_risk},
which are  passed to {simulate_risk}
}
\usage{
data(config_default_json)
}
\description{
JSON input configuration generated with R script (data-raw/config_json.R)
}
\examples{
data("config_default_json")
config_default_json

}
\keyword{datasets}
