% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{config_json}
\alias{config_json}
\title{config_json: configuration for kwb.qmra::opencpu_simulate_risk() in JSON format}
\format{
A json list with all configuration parameters used by \link{opencpu_simulate_risk},
which are  passed to {simulate_risk}
}
\usage{
data(config_json)
}
\description{
JSON input configuration generated with R script (data-raw/config_json.R)
}
\examples{
config_json

}
\keyword{datasets}
