% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_read}
\alias{config_read}
\title{Config: read configuration}
\usage{
config_read(confDir = system.file("extdata/configs/dummy", package =
  "kwb.qmra"))
}
\arguments{
\item{confDir}{directory to read configuration files (Default:
system.file('extdata/configs/dummy', package = 'kwb.qmra')")}
}
\value{
stores configuration in R list structure
}
\description{
Config: read configuration
}
