% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_write_dummy}
\alias{config_write_dummy}
\title{Config: create dummy configuration}
\usage{
config_write_dummy(
  confDir = system.file("extdata/configs/dummy", package = "kwb.qmra")
)
}
\arguments{
\item{confDir}{directory to save the dummy configuration files (Default:
system.file('extdata/config', package = 'kwb.qmra')")}
}
\value{
writes dummy configuration in confDir
}
\description{
Config: create dummy configuration
}
