% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{default_max}
\alias{default_max}
\title{Default Max}
\usage{
default_max(org_max, new_max, f = c)
}
\arguments{
\item{org_max}{original maximum value}

\item{new_max}{new maximum value for replacement}

\item{f}{function apply on "org_max" to transform to the correct 
dimension (e.g. "log" in case of "lognorm" or "log10" in case of log10_norm) 
(default: c())}
}
\description{
Default Max
}
\examples{
default_max(org_max = 2, new_max = 0.01, f = log10)
default_max(org_max = 0, new_max = 0.01, f = log10)
default_max(org_max = 2, new_max = 0.01, f = log)
default_max(org_max = 0, new_max = 0.01, f = log)
}
