% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{default_min}
\alias{default_min}
\title{Default Min}
\usage{
default_min(org_min, org_max, new_min, f = c)
}
\arguments{
\item{org_min}{original minimum value}

\item{org_max}{original maximum value}

\item{new_min}{new minimum value for replacement}

\item{f}{function apply on org_min and new_min to transform to the correct 
dimension (e.g. "log" in case of "lognorm" or "log10" in case of log10_norm),
(default: c())}
}
\description{
Default Min
}
\examples{
default_min(org_min = 2, org_max = 100, new_min = 0.01, f = log10)
default_min(org_min = 0, org_max = 100, new_min = 0.01, f = log10)
default_min(org_min = 2, org_max = 100, new_min = 0.01, f = log)
default_min(org_min = 0, org_max = 100, new_min = 0.01, f = log)
}
