% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{distribution_repeater}
\alias{distribution_repeater}
\title{Helper function: distributon repeater}
\usage{
distribution_repeater(number_of_repeatings = 10, number_of_events = 365,
  func, ...)
}
\arguments{
\item{number_of_repeatings}{how often should the random distribution with the 
same parameters be generated (default: 1)}

\item{number_of_events}{number of events}

\item{func}{distribution function to be repeated (e.g. runif, rlnorm, rnorm)}

\item{...}{further parameters passed to func}
}
\value{
data.frame with columns repeatID, eventID and values
}
\description{
Helper function: distributon repeater
}
\examples{
distribution_repeater(number_of_repeatings =  2,
                       number_of_events = 10,
                       func = runif,
                       min = 1,
                       max = 10)
}
