% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseresponse.R
\name{dr.db_download}
\alias{dr.db_download}
\title{Download dose-response model database from QMRAwiki}
\source{
http://qmrawiki.canr.msu.edu/index.php?title=Table_of_Recommended_Best-Fit_Parameters
}
\usage{
dr.db_download(fromInternet = FALSE)
}
\arguments{
\item{fromInternet}{download from internet (default: FALSE), if FALSE import from local copy}
}
\value{
tbl_df for different microbial parameters
}
\description{
Download dose-response model database from QMRAwiki
}
