% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseresponse.R
\name{dr.db_model}
\alias{dr.db_model}
\title{Generate table with different doses for dr.db_download()}
\usage{
dr.db_model(dr.db = dr.db_download(), dose = sfsmisc::lseq(from = 1, to
  = 10^10, length = 1000))
}
\arguments{
\item{dr.db}{as retrieved by dr.db_download(), default: dr.db_download()}

\item{dose}{vector of dose data (default: 
\code{sfsmisc::lseq(from=0.1, to = 10^10,length = 1000)})}
}
\value{
tbl_df
}
\description{
Generate table with different doses for dr.db_download()
}
