% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseresponse.R
\name{dr.expo}
\alias{dr.expo}
\title{Dose-response model: exponential}
\usage{
dr.expo(dose = sfsmisc::lseq(from = 1, to = 10^10, length = 1000), k = 0.572)
}
\arguments{
\item{dose}{vector of dose data (default: 
\code{sfsmisc::lseq(from = 0.1, to = 10^10, length = 1000)})}

\item{k}{k-value (default: 5.72E-01)}
}
\value{
tibble
}
\description{
Dose-response model: exponential
}
