% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{generate_random_values}
\alias{generate_random_values}
\title{Create random distribution based on configuration file}
\usage{
generate_random_values(config, number_of_repeatings = 1, number_of_events,
  debug = TRUE)
}
\arguments{
\item{config}{as retrieved by config_read()}

\item{number_of_repeatings}{how often should the random distribution with the 
same parameters be generated (default: 1)}

\item{number_of_events}{number of events}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list random distributions based on configuration file
}
\description{
Create random distribution based on configuration file
}

