% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{get_percentile}
\alias{get_percentile}
\title{Helper function: get percentile}
\usage{
get_percentile(percent_within_minmax = 0.9)
}
\arguments{
\item{percent_within_minmax}{percent of data point within min/max (default: 
0.9 i.e. 90 percent)}
}
\description{
Helper function: get percentile
}
\examples{
get_percentile(0.9)
get_percentile(0.95)

}
