% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_logremoval_stats.R
\name{get_risk_logremoval_stats_lean}
\alias{get_risk_logremoval_stats_lean}
\title{"Lean" version of get_risk_logremoval_stats()}
\usage{
get_risk_logremoval_stats_lean(data, config)
}
\description{
"Lean" version of get_risk_logremoval_stats()
}
