% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_total_stats.R
\name{get_risk_total_stats_lean}
\alias{get_risk_total_stats_lean}
\title{Lean version of get_risk_total_stats()}
\usage{
get_risk_total_stats_lean(data_lean, config)
}
\description{
This function returns the same as get_risk_total_stats() but does not require
the "Name" columns. Instead, names are merged by ID from the metadata tables
given in "config".
}
