% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{myggtheme}
\alias{myggtheme}
\title{Helper function: create default ggplot theme}
\usage{
myggtheme()
}
\value{
default ggplot theme for kwb.qmra
}
\description{
Helper function: create default ggplot theme
}
\keyword{internal}
