% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{number_of_exposures}
\alias{number_of_exposures}
\title{Helper function: gets number of exposures from config}
\usage{
number_of_exposures(config)
}
\arguments{
\item{config}{as retrieved by config_read()}
}
\value{
number of exposures
}
\description{
Helper function: gets number of exposures from config
}
\keyword{internal}
