% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{number_of_repeatings}
\alias{number_of_repeatings}
\title{Helper function: gets number of repeatings from config}
\usage{
number_of_repeatings(config)
}
\arguments{
\item{config}{as retrieved by config_read()}
}
\value{
number of repeatings (used for bayesian analysis)
}
\description{
Helper function: gets number of repeatings from config
}
\keyword{internal}
