% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{opencpu_config_read}
\alias{opencpu_config_read}
\title{OpenCPU wrapper: import config from CSV and convert to JSON}
\usage{
opencpu_config_read(
  confDir = system.file("extdata/configs/default", package = "kwb.qmra")
)
}
\arguments{
\item{confDir}{directory to read configuration files (Default:
system.file('extdata/configs/default', package = 'kwb.qmra')")}
}
\value{
stores configuration in JSON format
}
\description{
OpenCPU wrapper: import config from CSV and convert to JSON
}
\examples{
### Example json config file
config_json <- kwb.qmra::opencpu_config_read()
head(config_json)

}
