% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{opencpu_simulate_risk}
\alias{opencpu_simulate_risk}
\title{OpenCPU wrapper: run risk calculation and convert results to JSON format}
\usage{
opencpu_simulate_risk(
  config = config_read(),
  usePoisson = TRUE,
  debug = TRUE,
  lean = TRUE
)
}
\arguments{
\item{config}{config object as retrieved by \code{kwb.qmra::config_read()}}

\item{usePoisson}{should a poisson proccess (see function dose_perEvent()) be
used to calculate the dose_perEvent (TRUE) or just the exposure_perEvent 
column (FALSE), (default: TRUE)}

\item{debug}{print debug information (default: TRUE)}

\item{lean}{if \code{TRUE}, a "lean" version of this function is called, see
\code{kwb.qmra:::simulate_risk_lean}, (default: TRUE)}
}
\value{
JSON list with parameters of user defined random distribution and 
corresponding values
}
\description{
OpenCPU wrapper: run risk calculation and convert results to JSON format
}
\examples{
### Example simulation run
## Read from JSON
config_json <- kwb.qmra::opencpu_config_read()
config <- jsonlite::fromJSON(config_json)
## Optionally directly import from CSVs
# config <- kwb.qmra::config_read()
risk_json <- kwb.qmra::opencpu_simulate_risk(config)
writeLines(text = risk_json, "risk.json")

}
