% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_doseresponse}
\alias{plot_doseresponse}
\title{plot reduction}
\usage{
plot_doseresponse(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for reduction
}
\description{
plotting
}

