% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_effluent}
\alias{plot_effluent}
\title{plot effluent}
\usage{
plot_effluent(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for effluent
}
\description{
plotting
}
