% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_dalys}
\alias{plot_event_dalys}
\title{plot dalys_per_event}
\usage{
plot_event_dalys(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for dalys_per_event
}
\description{
plotting
}
