% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_dose}
\alias{plot_event_dose}
\title{plot dose per event}
\usage{
plot_event_dose(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for dose per event
}
\description{
plotting
}
