% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_exposure}
\alias{plot_event_exposure}
\title{plot exposure per event}
\usage{
plot_event_exposure(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for exposure per event
}
\description{
plotting
}
