% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_illnessProb}
\alias{plot_event_illnessProb}
\title{plot illness probability}
\usage{
plot_event_illnessProb(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for illness probability
}
\description{
plotting
}

