% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_infectionProb}
\alias{plot_event_infectionProb}
\title{plot infection probability}
\usage{
plot_event_infectionProb(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for infection probability
}
\description{
plotting
}

