% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_event_volume}
\alias{plot_event_volume}
\title{plot volume per event}
\usage{
plot_event_volume(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}
}
\value{
ggplot for volume per event
}
\description{
plotting volume
}
