% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_inflow}
\alias{plot_inflow}
\title{plot inflow}
\usage{
plot_inflow(risk)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk}
}
\value{
ggplot for inflow
}
\description{
plotting
}
