% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_total_dalys}
\alias{plot_total_dalys}
\title{plot total DALYs}
\usage{
plot_total_dalys(risk, labelling = FALSE, title = "", tolerance = 1e-06)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}

\item{labelling}{if TRUE labels with absolute DALYs will be plotted (default: 
FALSE)}

\item{title}{title for plot (default: "")}

\item{tolerance}{accecptable tolerance level of risk (default: 1E-6)}
}
\value{
ggplot for total DALYs
}
\description{
plotting
}

