% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_total_infectionProb}
\alias{plot_total_infectionProb}
\title{plot total infection probability}
\usage{
plot_total_infectionProb(risk, tolerance = 1e-04)
}
\arguments{
\item{risk}{list as retrieved by simulate_risk()}

\item{tolerance}{accecptable tolerance level of risk (default: 1E-4)}
}
\value{
ggplot for total infection probability
}
\description{
plotting
}

