% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{poisson_dose}
\alias{poisson_dose}
\title{Helper function: poisson distribution based on exposure per event}
\usage{
poisson_dose(exposure_perEvent)
}
\arguments{
\item{exposure_perEvent}{exposed organisms per event}
}
\value{
dose per event based on poisson process
}
\description{
Helper function: poisson distribution based on exposure per event
}

