% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_workflow}
\alias{report_workflow}
\title{Create report (not working for "shiny" reports)}
\usage{
report_workflow(confDirs = system.file("extdata/configs/", package =
  "kwb.qmra"), report_template_dir = system.file("extdata/report", package =
  "kwb.qmra"), report_template_name = "workflow.Rmd",
  report_output_dir = NULL, openReport = TRUE)
}
\arguments{
\item{confDirs}{directory containing subdirectory/ies of QMRA configurations
default: system.file("extdata/configs/", package = "kwb.qmra")}

\item{report_template_dir}{report template directory 
(default: system.file("extdata/report", package = "kwb.qmra"))}

\item{report_template_name}{default: "workflow.Rmd"}

\item{report_output_dir}{directory where report should be saved, if NULL 
report_template_dir is used (default: NULL)}

\item{openReport}{open report in browser default: TRUE}
}
\value{
generate html report
}
\description{
Create report (not working for "shiny" reports)
}

