% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{risk_json}
\alias{risk_json}
\title{risk_json: example risk_json object returned by kwb.qmra::opencpu_simulate_risk()}
\format{
A json list with all risk results generated by \link{opencpu_simulate_risk},
using the dataset `config_json`
}
\usage{
data(risk_json)
}
\description{
JSON risk result object generated with R script (data-raw/risk_json.R)
}
\examples{
data("risk_json")
risk <- jsonlite::fromJSON(risk_json)
### only show "stats" elements (skip "events" and "total" due to much data)
risk_stats <- list(stats_total = risk$stats_total,
                   stats_logremoval = risk$stats_logremoval)
jsonlite::toJSON(risk_stats, pretty = TRUE)
}
\keyword{datasets}
