% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_exposure}
\alias{simulate_exposure}
\title{Simulate: exposure}
\usage{
simulate_exposure(config, debug = TRUE)
}
\arguments{
\item{config}{as retrieved by config_read()}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list with parameters of user defined exposure scenario (number of events 
and volumes per Event)
}
\description{
Simulate: exposure
}

