% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{who_getTreatment}
\alias{who_getTreatment}
\title{Treatment: get WHO reductions}
\usage{
who_getTreatment()
}
\value{
list with $untidy and  $tidy data frames
}
\description{
Treatment: get WHO reductions
}
\seealso{
\url{http://apps.who.int/iris/bitstream/10665/44584/1/9789241548151_eng.pdf#page=162}
}
