% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ylogscale}
\alias{ylogscale}
\title{Helper function: prepare ylogscale appearence}
\usage{
ylogscale(paraVals, userLimit = NULL, by = 1)
}
\arguments{
\item{paraVals}{values to be used}

\item{userLimit}{vector specificing user limits to be used in case that paraVals 
are to not included}

\item{by}{by parameter for seq() (default: 1)}
}
\value{
default ggplot theme for kwb.qmra
}
\description{
Helper function: prepare ylogscale appearence
}
\keyword{internal}
