% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{create_random_distribution}
\alias{create_random_distribution}
\title{Create random distribution}
\usage{
create_random_distribution(type = "uniform", number_of_repeatings = 1,
  number_of_events = 365, value = 10, min = 10, max = 1000,
  mean = (min + max)/2, sdev = (max - mean)/qnorm(0.975),
  meanlog = mean(log((min + max)/2)), sdlog = sd(log((c(min, max)))),
  mode = (min + max)/2, debug = TRUE)
}
\arguments{
\item{type}{"uniform" calls runif(), "triangle" calls EnvStats::rtri(), "lognorm" 
calls rlnorm() and "norm" calls rnorm(), (default: "uniform")}

\item{number_of_repeatings}{how often should the random distribution with the 
same parameters be generated (default: 1)}

\item{number_of_events}{number of events}

\item{value}{constant value (no random number), gets repeated number_of_events
times (if 'type' = 'value')}

\item{min}{minimum value (default: 10), only used if 'type' is "runif" or 
"triangle"}

\item{max}{maximum value (default: 1000), only used if 'type' is "runif" or 
"triangle"}

\item{mean}{mean value (default: mean of min & max value), only used if 'type' 
is "norm"}

\item{sdev}{standard deviation (default: (max-mean) / qnorm(0.975)), 
only used if 'type' is "norm"}

\item{meanlog}{log mean value (default: mean of log min & max value), only 
used if 'type' is "lognorm"}

\item{sdlog}{standard deviation (default: sd of log min & max value), only 
used if 'type' is "lognorm"}

\item{mode}{(default: mean of min & max), only used if 'type' is "triangle"}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list with parameters of user defined random distribution and 
corresponding values
}
\description{
Create random distribution
}
\seealso{
for random triangle see \code{\link{rtri}}
}

