% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_distribution.R
\name{create_random_distribution}
\alias{create_random_distribution}
\title{Create random distribution}
\usage{
create_random_distribution(type = "uniform", number_of_repeatings = 1,
  number_of_events = 365, value = 10, min = 10, max = 1000,
  log10_zero_threshold = log10(0.01), log10_min = ifelse(min > 0 && max
  > 1, log10(min), ifelse(min == 0 && max < 1, 0.01 * max,
  log10_zero_threshold)), log10_max = ifelse(max > 0, log10(max),
  log10_zero_threshold), log10_mean = (log10_min + log10_max)/2,
  log10_sdev = (log10_max - log10_mean)/qnorm(0.9), mean = (min +
  max)/2, sdev = (max - mean)/qnorm(0.9), meanlog = mean(log((min +
  max)/2)), sdlog = sd(log((c(min, max)))), mode = (min + max)/2,
  debug = TRUE)
}
\arguments{
\item{type}{"uniform" calls runif(), "log10_uniform" calls 
10^runif(number_of_events, log10_min, log10_max), "triangle" calls 
EnvStats::rtri(), "lognorm" calls rlnorm(), "norm" calls rnorm() and 
"log10_norm" calls 10^rnorm(number_of_events, mean = log10_mean, 
sdev = log10_sdev), (default: "uniform")}

\item{number_of_repeatings}{how often should the random distribution with the
same parameters be generated (default: 1)}

\item{number_of_events}{number of events}

\item{value}{constant value (no random number), gets repeated number_of_events
times (if 'type' = 'value')}

\item{min}{minimum value (default: 10), only used if 'type' is "runif" or
"triangle"}

\item{max}{maximum value (default: 1000), only used if 'type' is "runif" or
"triangle"}

\item{log10_zero_threshold}{only used if 'type' is "log10_uniform" or 
"log10_norm" and "min" value equal zero. In this case the zero is replaced by 
this value (default: log10(0.01), i.e. with a lower limit of log10(0.01) = -2}

\item{log10_min}{minimum value (default: ifelse(min > 0 && max > 1, log10(min), 
ifelse(min == 0 && max < 1, 0.01 * max, log10_zero_threshold)), only used if 
'type' is "log10_uniform" or "log10_norm"}

\item{log10_max}{maximum value (default: ifelse(max > 0, log10(max), 
log10_zero_threshold), only used if 'type' is "log10_uniform" or "log10_norm"}

\item{log10_mean}{mean value (default: (log10_min + log10_max)/2), only used 
if 'type' is "log10_norm"}

\item{log10_sdev}{standard deviation (default: (log10_max- log10_mean) / 
qnorm(0.9)), only used if 'type' is "log10_norm"}

\item{mean}{mean value (default: mean of min & max value), only used if 'type'
is "norm"}

\item{sdev}{standard deviation (default: (max-mean) / qnorm(0.9)),
only used if 'type' is "norm"}

\item{meanlog}{log mean value (default: mean of log min & max value), only
used if 'type' is "lognorm"}

\item{sdlog}{standard deviation (default: sd of log min & max value), only
used if 'type' is "lognorm"}

\item{mode}{(default: mean of min & max), only used if 'type' is "triangle"}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list with parameters of user defined random distribution and
corresponding values
}
\description{
Create random distribution
}
\seealso{
for random triangle see \code{\link{rtri}}
}
