% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{opencpu_simulate_risk}
\alias{opencpu_simulate_risk}
\title{OpenCPU wrapper: Simulate risk using JSON format for config/results}
\usage{
opencpu_simulate_risk(
  config_json = opencpu_config_read(),
  usePoisson = TRUE,
  debug = TRUE,
  lean = TRUE
)
}
\arguments{
\item{config_json}{config json object as retrieved by opencpu_config_read()}

\item{usePoisson}{should a poisson proccess (see function dose_perEvent()) be
used to calculate the dose_perEvent (TRUE) or just the exposure_perEvent 
column (FALSE), (default: TRUE)}

\item{debug}{print debug information (default: TRUE)}

\item{lean}{if \code{TRUE}, a "lean" version of this function is called, see
\code{kwb.qmra:::simulate_risk_lean}, (default: TRUE)}
}
\value{
JSON list with parameters of user defined random distribution and 
corresponding values
}
\description{
OpenCPU wrapper: Simulate risk using JSON format for config/results
}
\examples{
### Example json config file
config_json <- kwb.qmra::opencpu_config_read()
risk_json <- kwb.qmra::opencpu_simulate_risk(config_json)

}
