% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_inflow}
\alias{simulate_inflow}
\title{Simulate: inflow}
\usage{
simulate_inflow(config, debug = TRUE)
}
\arguments{
\item{config}{as retrieved by config_read()}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list with parameters of user defined random distribution and 
corresponding values
}
\description{
Simulate: inflow
}
