% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_treatment}
\alias{simulate_treatment}
\title{Simulate: treatment}
\usage{
simulate_treatment(config, wide = FALSE, debug = TRUE)
}
\arguments{
\item{config}{as retrieved by config_read()}

\item{wide}{if TRUE results will be converted to wide format (default: 
FALSE)}

\item{debug}{print debug information (default: TRUE)}
}
\value{
list with parameters of user defined random distribution and 
corresponding values
}
\description{
Simulate: treatment
}
